+++ src/Ontology.h
@@ -49,9 +49,6 @@ public:
 private:
 
     const std::string ONTOLOGY_OUT_PATH     = "ontology/";
-    const std::string FINAL_ANNOT_FILE      = "final_annotations";
-    const std::string FINAL_ANNOT_FILE_CONTAM = "final_annotations_contam";
-    const std::string FINAL_ANNOT_FILE_NO_CONTAM = "final_annotations_no_contam";
 
     std::vector<ENTAP_HEADERS>      mEntapHeaders;
     std::vector<uint16>             mGoLevels;
@@ -60,14 +57,12 @@ private:
     std::string                     mOutpath;
     std::string                     mNewInput;
     std::string                     mOntologyDir;
-    std::string                     mFinalOutputDir;
     QueryData                       *mpQueryData;
     FileSystem                      *mpFileSystem;
     UserInput                       *mpUserInput;
     EntapDataPtrs                   mEntapDataPtrs;
     std::vector<FileSystem::ENT_FILE_TYPES> mAlignmentFileTypes;
 
-    void print_eggnog();
     std::unique_ptr<AbstractOntology> spawn_object(uint16&);
 };